fix: defer evaluation of $localize expressions to support runtime i18n for SSR
During SSR, we swap locales and translation catalogs on per-request basis, which means that we need to reevaluate $localize`foo` expressions on every request. This works today just fine as long as the $localize call is within a component body.

However if we have a module level constant that contains a JSX expression that is used later in a component template, the top level constant will be evaluated when the module initializer runs and not reevaluated on per request basis.

Unless addressed, this results in parts of the SSRed template to not honor the user locale, but instead use the default (often English) locale, while ignoring the translation catalogs.

This fix wraps all $localize calls into a labda which is reexecuted every time we render JSX to html, which ensures that even top level JSX "fragments" are localized using the per-request defined locale.